miniscript: register policy and check registration when receiving
The policy is registered just like a multisig account, by computing a
hash of the policy and storing it with a user-chosen account name.

The registration is checked before displaying a receive address.

The hash function does not include the origin and keypath of the key
origin info, as they don't influence the result: if e.g. a keypath
would change to one of our own keys, the xpub would change and so
would the hash. For keys that are not ours, we don't care about the
origin/keypath.