eth: add support for EIP-1559 transactions
https://github.com/ethereum/EIPs/blob/master/EIPS/eip-1559.md

EIP-1559 transactions start with a 0x02 byte followed by an RLP
encoding of the transaction fields. The chainID is newly included in
the fields directly instead of being encoded in the `v` value of the
signature.

We add a new proto message for EIP1559 signing as opposed to extending
the existing one for clarity - some fields don't exist
anymore (`coin`, `gas_price`) while new fields are
introduced (`max_priority_fee_per_gas` etc.).

`access_list` is presumed empty for the time being - we can add
support for a non-empty list if there is demand for it in the future.

The firmware verison is bumped to v9.16.0. The Python/Go/Rust client
libs assume that this is the version where EIP1559 is introduced.