show mnemonic: Disclaimer before recovery words
When the user wanted to see their recovery words, the only requirement
is unlocking the device and then confirming writing down the recovery
words. However, because of the scammers around like in Telegram, we
wanted to show an additional confirmation screen telling user not to
share their recovery words with anyone. Now it is implemented as another
step before showing the words and the user has to accept this step also.

Signed-off-by: asi345 <inanata15@gmail.com>