rust: allow restricting get_bip39_wordlist to subset
This will allow to set the autocomplete keyboard to a subset of the
BIP39 wordlist to allow entering the last word of a 12/18 word
mnemonic using the keyboard, similar to how the 24th word can be
entered using a menu.

The keyboard is needed for 12/18 words as there are too many
candidates to show in a simple menu.