simulator: make simulator the server side in socket connection
In the first implementation of socket connection in simulator, it was acting
as a client in the communication, trying to connect to a specific network host
and port. It was done like this because `send_message.py` running in host
machine was having problems connecting to a simulator

But actually, now that we do not need to connect BitBox02 device physically
while running the simulator, we also do not need to run `send_message.py` in
the host machine. Also ideally simulator should be a server hosting the
firmware to the clients.

Signed-off-by: asi345 <inanata15@gmail.com>