bip85: restructure api messages to allow for different bip85 apps
For now the same app as before is supported - bip39, but explicit.

The empty message now returns InvalidInput - would have been a
breaking change but the BIP85 API endpoint was never
released (disabled by not activating the app-bip85 Rust feature).

Reason: we want to add another dedicated app number for the generation
of deterministic entropy for the Breez SDK Lightning hot wallet in the
BitBoxApp. This will segregate keys so that BIP85-BIP39 mnemonics
which the user already might use (or will use) are not unintentionally
re-used and made hot in the Lightning wallet.