ci: disable address sanitizer for now
On my own machine inside podman, SANITIZE=ON leads to 100% cpu and
hanging unit tests. On a different computer it works fine.

GitHub actions used to work fine too until one of these days it
started to exhibit the same sympton - hangs forever and times out.

This happens also when re-running CI on old commits that previously
passed CI, so it's not a regression in the code.

I can't figure out what the cause of this is, but we a
reliable/passing CI, so we disable this check until we know how to fix
it.