simulator : multiple sequential client support
An issue that was faced while implementing HWI integration was
multiple client support. In HWI, one client could access to simulator for
initialization and then disconnect, while newer clients could connect
and directly test the functionalities of wallet API, assuming that the
device is already initialized. Parallel handling of clients are still
not supported, also not expected from HWI.

Signed-off-by: asi345 <inanata15@gmail.com>