Dockerfile: bump Rust toolchain to 1.78
This reduces the fw binary size by 4176 bytes as compared to the
v9.18.0 release.

Also bump LLVM 18 - it is unrelated but I did it anyway as the Rust
release notes also mentioned that they moved to LLVM 18.

The workflow.rs change is due to this deprecation warning:

```
warning: creating a shared reference to mutable static is discouraged
   --> bitbox02-rust-c/src/workflow.rs:103:11
    |
103 |     match &CONFIRM_STATE {
    |           ^^^^^^^^^^^^^^ shared reference to mutable static
    |
    = note: for more information, see issue #114447 <https://github.com/rust-lang/rust/issues/114447>
    = note: this will be a hard error in the 2024 edition
    = note: this shared reference has lifetime `'static`, but if the static ever gets mutated, or a mutable reference is created, then any further use of this shared reference is Undefined Behavior
    = note: `#[warn(static_mut_refs)]` on by default
help: use `addr_of!` instead to create a raw pointer
    |
103 |     match addr_of!(CONFIRM_STATE) {
```

The C code changes are clang-format changes. Clang-tidy
`-readability-avoid-unconditional-preprocessor-if` was added because
of the `#if 0` in pukcc.c, which seems worth preserving for now.