rust: fix vendored compilation of the git dep bip32-ed25519
Somehow, as bip32-ed25519 is a git dep, it was not actually compiled
from the local vendor folder but it was checked out into a git repo in
~/.cargo/git/ceckouts and compiled from there.

`cargo vendor` helpfully shows what needs to be added to the
config.toml so it is used.

We add the `--offline` flags to cargo build and cargo metadata to
prevent this from happening by accident in the future.