policies: refactor script derivation
Instead of deriving the witness script at a keypath, we instead derive
a descriptor instead.

The reason for this is that witness_script only makes sense for
`wsh(...)` desciptors. When we add `tr(...)` Taproot descriptors, we
will need the Taproot output key instead. So a `witness_script()`
method directly on the policy does not make sense as it does not work
for both variants.