Revert "policies: move duplicate key check into separate function"
This reverts commit 34bc408d4a0a8f95025d8d3292f5a8556ea969b0 (except
for typo).

BIP-388 prohibits duplicate keys across all keys in a Taproot
descriptor, even across the internal key and leaf scripts, so we
don't need to move this into a separate function (the reason for it
was so we could call it once per leaf script, which is not necessary).