btc/policies: check that keypath points to our key
When deriving the descriptor at the keypath. This is a basic sanity
check, as it does not make sense to derive it at the keypath used by a
cosigner, even though it results in the same descriptor. This check
also prevents signing inputs at a keypath belonging to a cosigner,
which does not seem dangerous in this case, but there is no point in
allowing it.