py/bitbox02: add boolean to get_info tuple resp.
The device now returns an additional byte from REQ_INFO: 0x00 if the
device is not initialized and 0x01 if it is. Read this byte in get_info
from the REQ_INFO response in case the firmware version supports it, and
add it to the tuple that is returned, None if the firmware does not
support it yet.