simulator: add --version argument
One purpose is to be able to track the current simulator version in
git history, so when we release it, we know which version it should
be.

bitbox02 prefix as we will have simulators for future BitBox versions
too.

The firmware version is there so we know which firmware is simulated.

The "simulator<VERSION>" part is to identify the simulator and its
version.

"linux-amd64" because we currently only support building using the
Dockerfile, with the output being linux-amd64.