transaction/fee: don't use fmt with "{.1}" to reduce binary size
Using this pulls in
`core::fmt::float::float_to_decimal_common_shortest` into the binary,
which is a huge function.

By changing the impl to something simpler, we save 20112 bytes in the
binary :O

We also add a CI check to avoid re-introducing this function in the
future by accident.