eth tx verification: use same address case as in input
Previously, no matter the format of the adress given by user to the
client, BitBox02 was showing this address in mixed case in the
verification step(see 'mixed-case checksum address encoding in EIP-55').

Even though this is acceptable, we thought that this could create
confusion among users when they see that the address they inputted is
not the same on the device(only difference is in the case though). To
avoid any questions, this commit achieves showing the address in the
same case as in user input on the device. Valid cases are : all upper,
all lower or mixed(EIP-55).

The client libraries will also implement a mechanism to categorize the
case of the input address to pass to BitBox02 firmware.

Signed-off-by: asi345 <inanata15@gmail.com>