ui_stub for simulator: show transaction address and fee confirm screens
Simulator was logging the normal confirm screens, but not showing the
Ethereum trransaction confirmation which were showing addresses and also
fees. Now, simulator logs them just like the normal confirmation screen
is shown.

This is useful since it provides completeness on simulator outputs and
also makes it easier to test/debug further changes.

Signed-off-by: asi345 <inanata15@gmail.com>