Script for vendoring dependencies of rust std libs
This commit creates a script `vendor.sh` that can be used to vendor our
dependencies from crates.io. In addition to vendoring the direct
dependencies of our rust crates, it also vendors the dependencies of the
rust standard libraries.

This script requires a minor update to the Dockerfile because we need to
install the sources of the rust standard libraries and copy a file when
we are the "root" user.