ci: Compare against HEAD instead of working tree
When we list modified files we should only check the commits that are
being added or are under review. Therefore we should use the format `git
diff <commit> <commit>`. Using `git diff <commit>` will diff <commit>
against the current working tree, which is only correct if the working
tree is pristine.

TARGET_BRANCH doesn't have to be limited to being a branch, it can be
any kind of git revision.

Finally instead of hard coding `master`, we use the `event.before` which
has the tip of the branch before the push event.