btc/paymentrequest: split memo into multiple screens
One screen to introduce the memo, and one screen per line in the
memo. This will allow a nice UX for Pocket to display:

"Memo from POCKET"

"IBAN correct?"

"CH 93 ...."

title_short vs title_long to not change the UX of signmsg while being
able to re-use the same workflow.

Having Pocket add the wording about "IBAN correct?" (or similar) is
a hack/workaround. It would be a better fit to modify SLIP-24 to add a
IBAN memo type, so the BitBox can add specific wording about this. We
don't do this for now to not deviate from the standard, but we might
in the future.