sdcard: auto-detection of insertion and delete 'remove sd card' feature
It seems that big share of the sd card support tickets are customers not clicking the upper right corner after inserting their sd card.
That's why auto-detection of sd card in the device is critical.

With this commit, when 'insert sd card' option is called from api,
BitBox02 device will start listening its sd card port and automatically
recognize if an sd card is inserted.

Furthermore, we noticed that 'remove sd card' feature is not used in any
of our libraries. That's why we removed it from api and Python library.

Signed-off-by: asi345 <inanata15@gmail.com>