Dockerfile: Maintenance
All temporary files need to be deleted in every `RUN` command to avoid adding it in a layer. Removing them in a later `RUN` command will still make them bloat the final image.

Instead of copying files temporarily they should be mounted in the `RUN`
command where they are needed. This avoids the need to delete them at
the end of the command.

Environment variables that are intended only for when the container is built should be prepended on the lines where they are needed to avoid polluting the environment in the running container.

We should strive to have as few layers as possible, but still enough to
simplify debugging the container image. When you only have 1 layer it is
not possible to create intermediate images and inspect the state of the
image.