Update to CMSIS 6.1
The latest version of CMSIS was fetched from the upstream github
repository[1]. It comes in a CMSIS-Pack format[2] (dear lord why...).
After some investigation it turns out that the CMSIS-Pack format is a
zip compressed folder with a custom extension. So the file was unzipped
and then the files that we obviously don't need (like support for other
cpus) was removed.

None of the breaking changes seems to have affected us and we get much
better support for the llvm/clang toolchain.

[1]: https://github.com/ARM-software/CMSIS_6/releases/tag/v6.1.0
[2]: https://www.open-cmsis-pack.org/