Add support for RTT through rust
This commit introduces a new `log` macro that calls `rprintln` in case
the `rtt` feature is enabled. It also creates a new c function called
`util_log` that reuses the rust RTT machinery.

The rust `log` macro is used the same way `println` is used. `util_log`
is used the same way `printf` is used.