sd: prompt user for formatting if format is not VFAT compatible
This commit inserts a check to validate whether the sdcard where the
backup will be put is VFAT compatible or not. VFAT compatilibity is
crucial for using the card and lack of it could result in errors in the
firmware. Therefore, when setting up the device, this commit validates
it and if it is not compatible, it asks user whether he/she is okay with
formatting or not. When the user confirms, SD card is formatted
accordingly and the fresh SD card is ready to be used by BitBox02.

Signed-off-by: asi345 <inanata15@gmail.com>