securechip: replace sign interface function with native Rust impl
The securechip unsafe sign key slot in the ATECC608 was only used as
an alternative to adding firmware code for signing using hte NIST
P-256 curve.

The Optiga Trust M does not support the same functionality - one can't
write a private key to this chip and extract the corresponding pubkey.

Since we want one interface for both chips, we remove the unsafe sign
functionality and replace it with the Rust p256 crate.

It is only used in U2F.

As an alternative, one could use the native MCU PUKCC/PUKCL feature to
derive a pubkey and sign using this curve (see pukcc.c), but I could
not get it to work. Using the Rust crate is much easier to implement /
use.

This adds 10592 bytes of binary space. If we need it back, we should
try again to use PUKCC instead and offload these operations to the
MCU.