rust: handle panics to show meaningful error
In b09c333 we added the `panic_immediate_abort` to the rust cargo flags.
This caused all the panics previously handled by the handler defined in
lib.rs to just show the `hard fault` error message on the screen of the
bitbox.

This adds the manual handling of some of those panics, to provide useful
debug messages.