src: change password functionality
This commit adds changing password feature. The feature can only be used
when the device is initialized.

It shares a lot of the functionality with set password feature, but it
requires the user to enter the old password correctly, and also does
some seed handling to not cause any data loss.

Signed-off-by: asi345 <inanata15@gmail.com>