add configuration for SPI_MEM
CLOCKS:
- SPI_MEM Core CLK: generic 0 (120MHz)
- SPI_MEM Slow CLK: generic 3 (48MHz)
- Sercom baud rate: 8MHz

The MX25 module needs 80MHz for high performance mode and 8MHz in low
power. Since the minimum divider for the baud rate generator is 4 (with
BAUD reg set to 1, see 33.6.2.3 of SAM D5 datasheet), we need an input
clock of 320MHz in high performance, which is not achievable with our
mcu. We'll go with the low-power mode.

I am not sure we actually need the low power clock.

PORTS:
Enabled SPI_MEM interface on SERCOM4 with the following pinout:
- CLK -> PA12
- MOSI -> PA15
- MISO -> PA13
- CS -> PA14
- WP -> PB07
- HOLD -> PA20

The code has been originally generated using the Atmel start tool and
adapted to the project.