bitcoin: display account name when sending to self
Before, we did this only when sending to the same account. This commit
extends this to work when sending to an address of a *different*
account of the same keystore.

The bitcoin transaction script_configs field is used for inputs and
change outputs. We add a similar list, `output_script_configs`, which
allows to attach script configs of different accounts to outputs, with
`output_script_config_index` to reference them in the output.

To ensure backwards compatibility, `output_script_config_index` is
added as a new field, which takes precedence over
`script_config_index` when set.

The output script configs are validated, but since one can send to any
address of the same keystore, the output script config validation
allows any combination of script configs, unlike the input script
configs. To not duplicate lots of code, a few refactors are done:

- validate_script_configs is now used for generic validation and is
used to validate the output script configs as well as the input script
configs
- validate_input_script_configs performs additional validation as
before (can't mix multisig and single sig inputs, can't mix inputs
from different accounts, etc).
- the multisig/policy name is stored in ValidatedScriptConfig, so we
don't have to fetch it twice

The Python lib version was downgraded from 8.0.0 to 7.0.0, it seems it
was bumped before in error - the previous release tag and release is 6.3.0.