usb: add product strings for bb02-plus, make dynamic
For the time being we intend to have one firmware per edition
targetting both BB02 and BB02-Plus. For this reason we add a memory
flag to store which platform it is (BB02 vs BB02-Plus), and make the
USB HID descriptor product string dynamic dependent on that.

Since our descriptor bytes are built from macros, this commit does a
quick solution by copy pasting the descriptor and changing only the
product string. In the future we can consider building the descriptor
bytes in normal code which would make setting the product string
easier.